Increase timeout for iam instance profile parameter
This increases the retry time on creating instances that fail due to
invalid instance profile values. This is typically an eventual
consistency problem. Even with the increased timeout, it's possible
that in certain time windows this would never succeed and just take
longer to fail. Increasing such timeouts have been quite successful
in the past though.